requirements.txt
certifi==2021.10.8
et-xmlfile==1.1.0
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
six==1.16.0
wincertstore==0.2
